Restructure directories for better organization
