+++ starterkits/drupal/app/_components/content/ArticleFull.tsx
