+++ starterkits/drupal/app/_components/content/BasicPageFull.tsx
