+++ starterkits/drupal/app/_components/navigation/MenuItemFragment.ts
