+++ starterkits/drupal/examples/_components/content/FilmTeaser.tsx
