+++ starterkits/drupal/examples/_components/layout/GlobalHeader.tsx
