+++ starterkits/drupal/examples/_components/navigation/SocialMenu.tsx
