+++ starterkits/drupal/examples/_components/paragraphs/QuoteParagraph.tsx
