+++ starterkits/drupal/examples/_components/paragraphs/TriviaParagraph.tsx
